requirements.txt
lxml==4.4.2
pkg-resources==0.0.0
python-docx==0.8.10
